Citizens Support - /trait chatbubble
This allows someone to set their Citizens based NPC to speak in ChatBubbles when speaking by doing "/trait chatbubble". Citizens support is only loaded when detected to avoid NPEs, and the trait is deregistered properly. Also, Citizens support may be disabled in Config if the player wanted to save resources. I had to rework the formatHologramLines methods slightly to allow LivingEntity instead of Player and expose some methods and variables to make it work. Check it out and thanks for making the plugin! I hope you add the feature so I can dump my other fork :P